adding new Anomaly Alert Template and making it work
