+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/anomaly/alert/util/AnomalyReportGenerator.java
@@ -6,10 +6,12 @@ import com.google.common.base.Throwables;
 import com.linkedin.thirdeye.anomaly.SmtpConfiguration;
 import com.linkedin.thirdeye.anomaly.ThirdEyeAnomalyConfiguration;
 import com.linkedin.thirdeye.anomaly.alert.AlertTaskRunner;
+import com.linkedin.thirdeye.anomaly.alert.v2.AlertTaskRunnerV2;
 import com.linkedin.thirdeye.client.DAORegistry;
 import com.linkedin.thirdeye.constant.AnomalyFeedbackType;
 import com.linkedin.thirdeye.datalayer.bao.MergedAnomalyResultManager;
 import com.linkedin.thirdeye.datalayer.bao.MetricConfigManager;
+import com.linkedin.thirdeye.datalayer.dto.AlertConfigDTO;
 import com.linkedin.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
 import com.linkedin.thirdeye.datalayer.dto.MetricConfigDTO;
 import freemarker.template.Configuration;
@@ -26,11 +28,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.apache.commons.mail.HtmlEmail;
+import org.joda.time.DateTimeZone;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class AnomalyReportGenerator {
-  // private static final DateFormat df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ");
 
   private static final Logger LOG = LoggerFactory.getLogger(AnomalyReportGenerator.class);
 
@@ -72,7 +74,8 @@ public class AnomalyReportGenerator {
   }
 
   public void buildReport(List<MergedAnomalyResultDTO> anomalies,
-      ThirdEyeAnomalyConfiguration configuration, String emailRecipients, String fromEmail) {
+      ThirdEyeAnomalyConfiguration configuration, AlertConfigDTO alertConfig) {
+    String subject = "Thirdeye report : " + alertConfig.getName();
     long startTime = System.currentTimeMillis();
     long endTime = 0;
     for (MergedAnomalyResultDTO anomaly : anomalies) {
@@ -83,12 +86,13 @@ public class AnomalyReportGenerator {
         endTime = anomaly.getEndTime();
       }
     }
-    buildReport(startTime, endTime, anomalies, configuration, false, emailRecipients, fromEmail);
+    buildReport(startTime, endTime, anomalies, subject, configuration, false,
+        alertConfig.getRecipients(), alertConfig.getFromAddress(), alertConfig.getName());
   }
 
   public void buildReport(long startTime, long endTime, List<MergedAnomalyResultDTO> anomalies,
-      ThirdEyeAnomalyConfiguration configuration, boolean includeSentAnomaliesOnly,
-      String emailRecipients, String fromEmail) {
+      String subject, ThirdEyeAnomalyConfiguration configuration, boolean includeSentAnomaliesOnly,
+      String emailRecipients, String fromEmail, String alertConfigName) {
     if (anomalies == null || anomalies.size() == 0) {
       LOG.info("No anomalies found to send email, please check the parameters.. exiting");
     } else {
@@ -138,6 +142,10 @@ public class AnomalyReportGenerator {
       }
 
       Map<String, Object> templateData = new HashMap<>();
+      DateTimeZone timeZone = DateTimeZone.forTimeZone(AlertTaskRunnerV2.DEFAULT_TIME_ZONE);
+      DataReportHelper.DateFormatMethod dateFormatMethod = new DataReportHelper.DateFormatMethod(timeZone);
+      templateData.put("timeZone", timeZone);
+      templateData.put("dateFormat", dateFormatMethod);
       templateData.put("startTime", new Date(startTime));
       templateData.put("endTime", new Date(endTime));
       templateData.put("anomalyCount", anomalies.size());
@@ -148,13 +156,16 @@ public class AnomalyReportGenerator {
       templateData.put("falseAlertCount", falseAlert);
       templateData.put("nonActionableCount", nonActionable);
       templateData.put("anomalyDetails", anomalyReportDTOList);
-      buildEmailTemplateAndSendAlert(templateData, configuration.getSmtpConfiguration(),
+      templateData.put("alertConfigName", alertConfigName);
+      templateData.put("reportGenerationTimeMillis", System.currentTimeMillis());
+      buildEmailTemplateAndSendAlert(templateData, configuration.getSmtpConfiguration(), subject,
           emailRecipients, fromEmail);
     }
   }
 
   void buildEmailTemplateAndSendAlert(Map<String, Object> paramMap,
-      SmtpConfiguration smtpConfiguration, String emailRecipients, String fromEmail) {
+      SmtpConfiguration smtpConfiguration, String subject, String emailRecipients,
+      String fromEmail) {
     if (Strings.isNullOrEmpty(fromEmail)) {
       throw new IllegalArgumentException("Invalid sender's email");
     }
@@ -168,11 +179,9 @@ public class AnomalyReportGenerator {
       Template template = freemarkerConfig.getTemplate("custom-anomaly-report.ftl");
       template.process(paramMap, out);
 
-      String alertEmailSubject = "Thirdeye : Daily anomaly report";
       String alertEmailHtml = new String(baos.toByteArray(), AlertTaskRunner.CHARSET);
-      EmailHelper
-          .sendEmailWithHtml(email, smtpConfiguration, alertEmailSubject, alertEmailHtml, fromEmail,
-              emailRecipients);
+      EmailHelper.sendEmailWithHtml(email, smtpConfiguration, subject, alertEmailHtml, fromEmail,
+          emailRecipients);
     } catch (Exception e) {
       Throwables.propagate(e);
     }
@@ -195,7 +204,8 @@ public class AnomalyReportGenerator {
   }
 
   String getAnomalyURL(MergedAnomalyResultDTO anomalyResultDTO, String dashboardUrl) {
-    String urlPart = "#view=anomalies&dataset=%s&metrics=%s&currentStart=%s&currentEnd=%s";
+    String urlPart =
+        "/dashboard#view=anomalies&dataset=%s&metrics=%s&currentStart=%s&currentEnd=%s";
     return dashboardUrl + String
         .format(urlPart, anomalyResultDTO.getCollection(), anomalyResultDTO.getMetric(),
             anomalyResultDTO.getStartTime(), anomalyResultDTO.getEndTime());
